remove: tiny sankeykey modules
