--- src/pages/Demos/Sankeykey/Legend.tsx
@@ -1,34 +1,0 @@
-import { Box, HStack, Text } from "@chakra-ui/react";
-import { useAtomValue } from "jotai";
-import {
-  CATEGORY_LABELS,
-  OUTCOME_ORDER,
-  categoryToken,
-} from "../TrajectoryAtlas/taxonomy";
-import { legendCategoriesAtom } from "./atoms";
-import { OUTCOME_PALETTE_KEY } from "./outcomeColors";
-
-function Chip({ token, label }: { token: string; label: string }) {
-  return (
-    <HStack gap={1.5}>
-      <Box w="10px" h="10px" borderRadius="sm" bg={token} />
-      <Text>{label}</Text>
-    </HStack>
-  );
-}
-
-export function Legend() {
-  const categories = useAtomValue(legendCategoriesAtom);
-
-  return (
-    <HStack flexWrap="wrap" gap={3} fontSize="xs" color="fg.muted">
-      {categories.map((cat) => (
-        <Chip key={cat} token={categoryToken(cat)} label={CATEGORY_LABELS[cat]} />
-      ))}
-      <Box w="1px" alignSelf="stretch" bg="gray.subtle" />
-      {OUTCOME_ORDER.map((o) => (
-        <Chip key={o} token={`chart.${OUTCOME_PALETTE_KEY[o]}`} label={o} />
-      ))}
-    </HStack>
-  );
-}
